+++ src/lib/model.spec.ts
@@ -1,6 +1,11 @@
-import {expect} from 'chai'
+import {assert, expect} from 'chai'
+import {Course, Module, RequiredRecurringAudience} from 'lib/model'
 import * as model from 'lib/model'
 import {AreaOfWork, Grade} from 'lib/registry'
+import {CourseRecord} from 'lib/service/learnerRecordAPI/courseRecord/models/courseRecord'
+import {RecordState} from 'lib/service/learnerRecordAPI/models/record'
+import {ModuleRecord} from 'lib/service/learnerRecordAPI/moduleRecord/models/moduleRecord'
+import * as sinon from 'sinon'
 
 const genericAOW = new AreaOfWork(1, 'co')
 const genericGrade = new Grade('Test', 'Test')
@@ -80,5 +85,133 @@ describe('Should test User roles logic', () => {
 		user.grade = genericGrade
 
 		expect(user.hasRole('learner')).to.equal(false)
+	})
+})
+
+const getBasicCourseRecord = () => {
+	return new CourseRecord('Test001', '', RecordState.Null, [], '', false)
+}
+
+const getBasicModuleRecord = (moduleId: string, state: RecordState, lastUpdated: Date, completedDate?: Date) => {
+	const moduleRecord = new ModuleRecord(0, moduleId, '', '', new Date(),
+		lastUpdated, '', 'link', state, 0, false, undefined)
+	moduleRecord.completionDate = completedDate
+	return moduleRecord
+}
+
+describe('displayState tests', () => {
+	/**
+	 * Audience with a due by date of 31/03/2020 on a yearly frequency
+	 */
+	const audience = new RequiredRecurringAudience(new Date("2020-03-31"), new Date("2021-03-31"))
+	const course = new Course("testId")
+	const module1 = Module.create({id: "1", type: "elearning", optional: true})
+	const module2 = Module.create({id: "2", type: "elearning", optional: false})
+	const module3 = Module.create({id: "3", type: "elearning", optional: false})
+	course.modules = [module1, module2, module3]
+
+	const lastYearDate = new Date("2020-03-30")
+	const thisYearDate = new Date("2020-04-05")
+	describe('getDisplayStateForCourse tests', () => {
+		const sandbox = sinon.createSandbox()
+		sandbox.stub(course, 'getRequiredRecurringAudience').returns(audience)
+		const courseRecord = getBasicCourseRecord()
+		courseRecord.modules = [
+			getBasicModuleRecord("1", RecordState.Completed, lastYearDate, lastYearDate),
+			getBasicModuleRecord("2", RecordState.Completed, lastYearDate, lastYearDate),
+			getBasicModuleRecord("3", RecordState.Completed, lastYearDate, lastYearDate),
+		]
+
+		describe("When the course has already been completed", () => {
+			it('should set the state to not started (null) when all modules have' +
+				'been completed during the previous learning year', () => {
+				courseRecord.getModuleRecord("1")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("2")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("3")!.completionDate = lastYearDate
+				const result = course.getDisplayState(courseRecord)
+				assert.equal(result, RecordState.Null)
+			})
+			it('should set the state to not started (null) when all required modules have' +
+				'been completed during the previous learning year', () => {
+				courseRecord.getModuleRecord("1")!.completionDate = thisYearDate
+				courseRecord.getModuleRecord("2")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("3")!.completionDate = lastYearDate
+				const result = course.getDisplayState(courseRecord)
+				assert.equal(result, RecordState.Null)
+			})
+			it('should set the state to not started (null) when required modules have' +
+				'been completed during the previous learning year and an optional module' +
+				'is in-progress this year', () => {
+				courseRecord.getModuleRecord("1")!.completionDate = thisYearDate
+				courseRecord.getModuleRecord("2")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("3")!.completionDate = lastYearDate
+				const result = course.getDisplayState(courseRecord)
+				assert.equal(result, RecordState.Null)
+			})
+			it('should set the state to in-progress when a module has ' +
+				'been completed during the previous learning year and another' +
+				'required module has been completed in the current learning year', () => {
+				courseRecord.getModuleRecord("1")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("2")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("3")!.completionDate = thisYearDate
+				const result = course.getDisplayState(courseRecord)
+				assert.equal(result, RecordState.InProgress)
+			})
+			it('should set the state to completed one required module has' +
+				'been progressed during the current learning year', () => {
+				courseRecord.getModuleRecord("1")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("2")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("2")!.updatedAt = thisYearDate
+				courseRecord.getModuleRecord("3")!.completionDate = lastYearDate
+				const result = course.getDisplayState(courseRecord)
+				assert.equal(result, RecordState.InProgress)
+			})
+			it('should set the state to completed when both required modules have' +
+				'been completed during the current learning year', () => {
+				courseRecord.getModuleRecord("1")!.completionDate = lastYearDate
+				courseRecord.getModuleRecord("2")!.completionDate = thisYearDate
+				courseRecord.getModuleRecord("3")!.completionDate = thisYearDate
+				const result = course.getDisplayState(courseRecord)
+				assert.equal(result, RecordState.Completed)
+			})
+		})
+
+	})
+
+	describe('getDisplayStateForModule tests', () => {
+
+		it('should set the state to not started (null) when a module is' +
+			'completed during the previous learning year', () => {
+			const moduleRecord = getBasicModuleRecord("1", RecordState.Completed, lastYearDate, lastYearDate)
+			const result = module1.getDisplayState(moduleRecord, audience)
+			assert.equal(result, null)
+		})
+		it('should set the state to not started (null) when a module is' +
+			'in progress during the previous learning year', () => {
+			const moduleRecord = getBasicModuleRecord("1", RecordState.InProgress, lastYearDate)
+			const result = module1.getDisplayState(moduleRecord, audience)
+			assert.equal(result, null)
+		})
+		it('should set the state to IN_PROGRESS when a module is' +
+			'in progress during the current learning year', () => {
+			const moduleRecord = getBasicModuleRecord("1", RecordState.InProgress, thisYearDate)
+			const result = module1.getDisplayState(moduleRecord, audience)
+			assert.equal(result, "IN_PROGRESS")
+		})
+		it('should set the state to COMPLETED when a module is' +
+			'completed during the current learning year', () => {
+			const moduleRecord = getBasicModuleRecord("1", RecordState.Completed, thisYearDate, thisYearDate)
+			const result = module1.getDisplayState(moduleRecord, audience)
+			assert.equal(result, "COMPLETED")
+		})
+		it('should set the state to COMPLETED when a module has been completed on a non-recurring course ', () => {
+			const moduleRecord = getBasicModuleRecord("1", RecordState.Completed, lastYearDate, lastYearDate)
+			const result = module1.getDisplayState(moduleRecord, null)
+			assert.equal(result, "COMPLETED")
+		})
+		it('should set the state to NULL when a module has no record associated ', () => {
+			const result = module1.getDisplayState(null, null)
+			assert.equal(result, null)
+		})
 	})
 })
